+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/notifier_transports.php
@@ -96,7 +96,7 @@ return static function (ContainerConfigurator $container) {
             ->parent('notifier.transport_factory.abstract')
             ->tag('texter.transport_factory')
 
-        ->set('notifier.transport_factory.all-my-sms', AllMySmsTransportFactory::class)
+        ->set('notifier.transport_factory.allmysms', AllMySmsTransportFactory::class)
             ->parent('notifier.transport_factory.abstract')
             ->tag('texter.transport_factory')
 
@@ -168,7 +168,7 @@ return static function (ContainerConfigurator $container) {
             ->parent('notifier.transport_factory.abstract')
             ->tag('chatter.transport_factory')
 
-        ->set('notifier.transport_factory.gateway-api', GatewayApiTransportFactory::class)
+        ->set('notifier.transport_factory.gatewayapi', GatewayApiTransportFactory::class)
             ->parent('notifier.transport_factory.abstract')
             ->tag('texter.transport_factory')