+++ src/Symfony/Component/Cache/Adapter/TagAwareAdapter.php
@@ -41,7 +41,7 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
     private static \Closure $createCacheItem;
     private static \Closure $setCacheItemTags;
     private static \Closure $getTagsByKey;
-    private static \Closure $invalidateTags;
+    private static \Closure $saveTags;
 
     public function __construct(AdapterInterface $itemsPool, AdapterInterface $tagsPool = null, float $knownTagVersionsTtl = 0.15)
     {
@@ -95,8 +95,10 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
             null,
             CacheItem::class
         );
-        self::$invalidateTags ?? self::$invalidateTags = \Closure::bind(
+        self::$saveTags ?? self::$saveTags = \Closure::bind(
             static function (AdapterInterface $tagsAdapter, array $tags) {
+                ksort($tags);
+
                 foreach ($tags as $v) {
                     $v->expiry = 0;
                     $tagsAdapter->saveDeferred($v);
@@ -114,40 +116,14 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
      */
     public function invalidateTags(array $tags): bool
     {
-        $ok = true;
-        $tagsByKey = [];
-        $invalidatedTags = [];
+        $ids = [];
         foreach ($tags as $tag) {
             \assert('' !== CacheItem::validateKey($tag));
-            $invalidatedTags[$tag] = 0;
+            unset($this->knownTagVersions[$tag]);
+            $ids[] = $tag.static::TAGS_PREFIX;
         }
 
-        if ($this->deferred) {
-            $items = $this->deferred;
-            foreach ($items as $key => $item) {
-                if (!$this->pool->saveDeferred($item)) {
-                    unset($this->deferred[$key]);
-                    $ok = false;
-                }
-            }
-
-            $tagsByKey = (self::$getTagsByKey)($items);
-            $this->deferred = [];
-        }
-
-        $tagVersions = $this->getTagVersions($tagsByKey, $invalidatedTags);
-        $f = self::$createCacheItem;
-
-        foreach ($tagsByKey as $key => $tags) {
-            $this->pool->saveDeferred($f(static::TAGS_PREFIX.$key, array_intersect_key($tagVersions, $tags), $items[$key]));
-        }
-        $ok = $this->pool->commit() && $ok;
-
-        if ($invalidatedTags) {
-            $ok = (self::$invalidateTags)($this->tags, $invalidatedTags) && $ok;
-        }
-
-        return $ok;
+        return !$tags || $this->tags->deleteItems($ids);
     }
 
     /**
@@ -174,7 +150,7 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
         }
 
         foreach ($this->getTagVersions([$itemTags]) as $tag => $version) {
-            if ($itemTags[$tag] !== $version && 1 !== $itemTags[$tag] - $version) {
+            if ($itemTags[$tag] !== $version) {
                 return false;
             }
         }
@@ -298,7 +274,30 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
      */
     public function commit(): bool
     {
-        return $this->invalidateTags([]);
+        if (!$this->deferred) {
+            return true;
+        }
+
+        $ok = true;
+        foreach ($this->deferred as $key => $item) {
+            if (!$this->pool->saveDeferred($item)) {
+                unset($this->deferred[$key]);
+                $ok = false;
+            }
+        }
+
+        $items = $this->deferred;
+        $tagsByKey = (self::$getTagsByKey)($items);
+        $this->deferred = [];
+
+        $tagVersions = $this->getTagVersions($tagsByKey);
+        $f = self::$createCacheItem;
+
+        foreach ($tagsByKey as $key => $tags) {
+            $this->pool->saveDeferred($f(static::TAGS_PREFIX.$key, array_intersect_key($tagVersions, $tags), $items[$key]));
+        }
+
+        return $this->pool->commit() && $ok;
     }
 
     public function __sleep(): array
@@ -342,7 +341,7 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
 
                 foreach ($itemTags as $key => $tags) {
                     foreach ($tags as $tag => $version) {
-                        if ($tagVersions[$tag] !== $version && 1 !== $version - $tagVersions[$tag]) {
+                        if ($tagVersions[$tag] !== $version) {
                             unset($itemTags[$key]);
                             continue 2;
                         }
@@ -358,42 +357,32 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
         }
     }
 
-    private function getTagVersions(array $tagsByKey, array &$invalidatedTags = []): array
+    private function getTagVersions(array $tagsByKey): array
     {
-        $tagVersions = $invalidatedTags;
+        $tagVersions = [];
+        $fetchTagVersions = false;
 
         foreach ($tagsByKey as $tags) {
             $tagVersions += $tags;
+
+            foreach ($tags as $tag => $version) {
+                if ($tagVersions[$tag] !== $version) {
+                    unset($this->knownTagVersions[$tag]);
+                }
+            }
         }
 
         if (!$tagVersions) {
             return [];
         }
 
-        if (!$fetchTagVersions = 1 !== \func_num_args()) {
-            foreach ($tagsByKey as $tags) {
-                foreach ($tags as $tag => $version) {
-                    if ($tagVersions[$tag] > $version) {
-                        $tagVersions[$tag] = $version;
-                    }
-                }
-            }
-        }
-
         $now = microtime(true);
         $tags = [];
         foreach ($tagVersions as $tag => $version) {
             $tags[$tag.static::TAGS_PREFIX] = $tag;
-            if ($fetchTagVersions || !isset($this->knownTagVersions[$tag])) {
+            if ($fetchTagVersions || ($this->knownTagVersions[$tag][1] ?? null) !== $version || $now - $this->knownTagVersions[$tag][0] >= $this->knownTagVersionsTtl) {
+                // reuse previously fetched tag versions up to the ttl
                 $fetchTagVersions = true;
-                continue;
-            }
-            $version -= $this->knownTagVersions[$tag][1];
-            if ((0 !== $version && 1 !== $version) || $now - $this->knownTagVersions[$tag][0] >= $this->knownTagVersionsTtl) {
-                // reuse previously fetched tag versions up to the ttl, unless we are storing items or a potential miss arises
-                $fetchTagVersions = true;
-            } else {
-                $this->knownTagVersions[$tag][1] += $version;
             }
         }
 
@@ -401,12 +390,18 @@ class TagAwareAdapter implements TagAwareAdapterInterface, TagAwareCacheInterfac
             return $tagVersions;
         }
 
+        $newTags = [];
+        $newVersion = null;
         foreach ($this->tags->getItems(array_keys($tags)) as $tag => $version) {
-            $tagVersions[$tag = $tags[$tag]] = $version->get() ?: 0;
-            if (isset($invalidatedTags[$tag])) {
-                $invalidatedTags[$tag] = $version->set(++$tagVersions[$tag]);
+            if (!$version->isHit()) {
+                $newTags[$tag] = $version->set($newVersion ?? $newVersion = random_int(\PHP_INT_MIN, \PHP_INT_MAX));
             }
+            $tagVersions[$tag = $tags[$tag]] = $version->get();
             $this->knownTagVersions[$tag] = [$now, $tagVersions[$tag]];
+        }
+
+        if ($newTags) {
+            (self::$saveTags)($this->tags, $newTags);
         }
 
         return $tagVersions;