+++ src/Symfony/Component/Cache/LockRegistry.php
@@ -42,7 +42,6 @@ final class LockRegistry
         __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'CouchbaseBucketAdapter.php',
         __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'CouchbaseCollectionAdapter.php',
         __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'DoctrineDbalAdapter.php',
-        __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'DoctrineSchemaConfiguratorInterface.php',
         __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'FilesystemAdapter.php',
         __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'FilesystemTagAwareAdapter.php',
         __DIR__.\DIRECTORY_SEPARATOR.'Adapter'.\DIRECTORY_SEPARATOR.'MemcachedAdapter.php',
@@ -101,6 +100,7 @@ final class LockRegistry
 
         while (true) {
             try {
+                $locked = false;
                 // race to get the lock in non-blocking mode
                 if ($wouldBlock = \function_exists('sem_get')) {
                     $locked = @sem_acquire($lock, true);
@@ -136,10 +136,12 @@ final class LockRegistry
                     flock($lock, \LOCK_SH);
                 }
             } finally {
-                if (\function_exists('sem_get')) {
-                    sem_remove($lock);
-                } else {
-                    flock($lock, \LOCK_UN);
+                if ($locked) {
+                    if (\function_exists('sem_get')) {
+                        sem_remove($lock);
+                    } else {
+                        flock($lock, \LOCK_UN);
+                    }
                 }
                 unset(self::$lockedKeys[$key]);
             }