+++ src/Symfony/Component/Mime/Part/DataPart.php
@@ -20,6 +20,9 @@ use Symfony\Component\Mime\MimeTypes;
  */
 class DataPart extends TextPart
 {
+    /** @internal */
+    protected $_parent;
+
     private static $mimeTypes;
 
     private $filename;
@@ -32,6 +35,8 @@ class DataPart extends TextPart
      */
     public function __construct($body, string $filename = null, string $contentType = null, string $encoding = null)
     {
+        unset($this->_parent);
+
         if (null === $contentType) {
             $contentType = 'application/octet-stream';
         }