+++ src/Symfony/Component/Mime/Part/SMimePart.php
@@ -18,6 +18,9 @@ use Symfony\Component\Mime\Header\Headers;
  */
 class SMimePart extends AbstractPart
 {
+    /** @internal */
+    protected $_headers;
+
     private $body;
     private $type;
     private $subtype;
@@ -25,6 +28,8 @@ class SMimePart extends AbstractPart
 
     public function __construct(iterable|string $body, string $type, string $subtype, array $parameters)
     {
+        unset($this->_headers);
+
         parent::__construct();
 
         $this->body = $body;