+++ src/Symfony/Component/Security/Core/Tests/Exception/CustomUserMessageAuthenticationExceptionTest.php
@@ -18,6 +18,8 @@ use Symfony\Component\Security\Core\User\InMemoryUser;
 
 class ChildCustomUserMessageAuthenticationException extends CustomUserMessageAuthenticationException
 {
+    public $childMember;
+
     public function __serialize(): array
     {
         return [$this->childMember, parent::__serialize()];